// src/app/modules/transportation/models/inspection-history-filter-view-model.ts
import { PagingParams } from 'app/models/paging-params';
import { SortOrder } from 'app/models/sort-order';

/**
 * Transport
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * InspectionHistoryViewModel
 */
export interface InspectionHistoryFilterViewModel extends PagingParams {
   /**
 * Gets Or Sets
 */
   vehicleId?: string;
   /**
    * Gets Or Sets
    */
   vehicleName?: string;
   /**
* Gets Or Sets
*/
   registrationNumber?: string;
   /**
    * Gets Or Sets
    */
   inspectionStartDateBegin?: Date;
   /**
* Gets Or Sets
*/
   inspectionStartDateEnd?: Date;
   /**
    * Gets Or Sets
    */
   odometerReadingAtInspection?: string;
   /**
    * Gets Or Sets
    */
   odometerReadingNextInspection?: string;
   /**
    * Gets Or Sets
    */
   nextInspectionStartDateBegin?: Date;
   /**
* Gets Or Sets
*/
   nextInspectionStartDateEnd?: Date;

   /**Gets or Sets sortOrder */
   sortOrder: SortOrder;

   /**Gets or Sets sortBy- */
   sortBy: string;
}