// src/app/modules/transportation/models/inspection-history-view.ts
import { AttachmentViewModel } from 'app/modules/staff/staff-admissions/model/attachment-view-model';

/**
 * Transport
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * 
 */
export interface InspectionHistoryView {
    /**
     * Gets Or Sets
     */
    vehicleId?: string;
    /**
     * Gets Or Sets
     */
    inspectionDate?: Date;
    /**
     * Gets Or Sets
     */
    odometerReadingAtInspection?: number;
    /**
     * Gets Or Sets
     */
    odometerReadingNextInspection?: number;
    /**
     * Gets Or Sets
     */
    nextInspectionDate?: Date;
    /**
     * Gets Or Sets
     */
    attachment?: AttachmentViewModel;
    /**
 * Gets Or Sets
 */
    id?: string;

}